Rename Cache files to AsyncDispatcher
